sabudh-agro-evidence/requirements.txt
celery
werkzeug
boto3
sqlalchemy
mysql-connector-python-rf
flask
pandas
pyjwt
passlib